Fix compile error

+++ src/Network/ExecutorFactory.cpp
@@ -5,23 +5,22 @@ using namespace Network;
 ExecutorFactory::ExecutorFactory() {
 
     executorList = QMap<quint8, AbstractExecutor*>();
-
-    executorList.insert(UsersExecutor::INSTRUCTION_NUMBER, new UsersExecutor);
-    executorList.insert(MoveNodeExecutor::INSTRUCTION_NUMBER, new MoveNodeExecutor);
-    executorList.insert(GraphStartExecutor::INSTRUCTION_NUMBER, new GraphStartExecutor);
-    executorList.insert(GraphEndExecutor::INSTRUCTION_NUMBER, new GraphEndExecutor);
-    executorList.insert(NewNodeExecutor::INSTRUCTION_NUMBER, new NewNodeExecutor);
-    executorList.insert(NewEdgeExecutor::INSTRUCTION_NUMBER, new NewEdgeExecutor);
-    executorList.insert(LayoutExecutor::INSTRUCTION_NUMBER, new LayoutExecutor);
-    executorList.insert(MoveAvatarExecutor::INSTRUCTION_NUMBER, new MoveAvatarExecutor);
-    executorList.insert(IncommingUserExecutor::INSTRUCTION_NUMBER, new IncommingUserExecutor);
-    executorList.insert(MessageExecutor::INSTRUCTION_NUMBER, new MessageExecutor);
-    executorList.insert(ServerStopExecutor::INSTRUCTION_NUMBER, new ServerStopExecutor);
-    executorList.insert(WelcomeExecutor::INSTRUCTION_NUMBER, new WelcomeExecutor);
-    executorList.insert(SendGraphExecutor::INSTRUCTION_NUMBER, new SendGraphExecutor);
-    executorList.insert(SpyUserExecutor::INSTRUCTION_NUMBER, new SpyUserExecutor);
-    executorList.insert(UnspyUserExecutor::INSTRUCTION_NUMBER, new UnspyUserExecutor);
-    executorList.insert(RemoveNodeExecutor::INSTRUCTION_NUMBER, new RemoveNodeExecutor);
+    executorList.insert(quint8(UsersExecutor::INSTRUCTION_NUMBER), new UsersExecutor);
+    executorList.insert(quint8(MoveNodeExecutor::INSTRUCTION_NUMBER), new MoveNodeExecutor);
+    executorList.insert(quint8(GraphStartExecutor::INSTRUCTION_NUMBER), new GraphStartExecutor);
+    executorList.insert(quint8(GraphEndExecutor::INSTRUCTION_NUMBER), new GraphEndExecutor);
+    executorList.insert(quint8(NewNodeExecutor::INSTRUCTION_NUMBER), new NewNodeExecutor);
+    executorList.insert(quint8(NewEdgeExecutor::INSTRUCTION_NUMBER), new NewEdgeExecutor);
+    executorList.insert(quint8(LayoutExecutor::INSTRUCTION_NUMBER), new LayoutExecutor);
+    executorList.insert(quint8(MoveAvatarExecutor::INSTRUCTION_NUMBER), new MoveAvatarExecutor);
+    executorList.insert(quint8(IncommingUserExecutor::INSTRUCTION_NUMBER), new IncommingUserExecutor);
+    executorList.insert(quint8(MessageExecutor::INSTRUCTION_NUMBER), new MessageExecutor);
+    executorList.insert(quint8(ServerStopExecutor::INSTRUCTION_NUMBER), new ServerStopExecutor);
+    executorList.insert(quint8(WelcomeExecutor::INSTRUCTION_NUMBER), new WelcomeExecutor);
+    executorList.insert(quint8(SendGraphExecutor::INSTRUCTION_NUMBER), new SendGraphExecutor);
+    executorList.insert(quint8(SpyUserExecutor::INSTRUCTION_NUMBER), new SpyUserExecutor);
+    executorList.insert(quint8(UnspyUserExecutor::INSTRUCTION_NUMBER), new UnspyUserExecutor);
+    executorList.insert(quint8(RemoveNodeExecutor::INSTRUCTION_NUMBER), new RemoveNodeExecutor);
 
 }
